requirements.txt
Click==7.0
dateutils==0.6.6
Flask==1.0.3
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
Werkzeug==0.15.4
